Refactor file and asset search into separate services
