+++ apps/server-asset-sg/src/features/assets/search/asset-search.service.ts
@@ -1,5 +1,4 @@
 import {
-  Asset,
   AssetId,
   AssetJSON,
   AssetSearchQuery,
@@ -9,9 +8,6 @@ import {
   AssetSearchStats,
   AssetSearchUsageCode,
   ContactId,
-  ElasticsearchAsset,
-  FileSearchResult,
-  FileSearchResultItem,
   GeometryType,
   LocalDate,
   User,
@@ -21,183 +17,28 @@ import {
 import { Client as ElasticsearchClient } from '@elastic/elasticsearch';
 import {
   AggregationsAggregationContainer,
-  QueryDslDateRangeQuery,
   QueryDslQueryContainer,
   SearchResponse,
   SearchTotalHits,
 } from '@elastic/elasticsearch/lib/api/types';
-import { Injectable, Logger } from '@nestjs/common';
-
+import { Injectable } from '@nestjs/common';
 import { WorkflowStatus } from '@swissgeol/ui-core';
 import { plainToInstance } from 'class-transformer';
+import { ASSET_ELASTIC_INDEX, SEARCH_BATCH_SIZE } from '@/features/assets/search/asset-search.constants';
+import {
+  mapQueryToElasticDsl,
+  mapQueryToElasticDslParts,
+  PageOptions,
+  SearchState,
+} from '@/features/assets/search/asset-search.query';
 
-// eslint-disable-next-line @nx/enforce-module-boundaries
-import indexMapping from '../../../../../../development/init/elasticsearch/mappings/swissgeol_asset_asset.json';
-// eslint-disable-next-line @nx/enforce-module-boundaries
-import fileIndexMapping from '../../../../../../development/init/elasticsearch/mappings/swissgeol_asset_file.json';
-
-import { PrismaService } from '@/core/prisma.service';
-import { AssetRepo } from '@/features/assets/asset.repo';
-import { mapLv95ToElastic } from '@/features/assets/search/asset-search.utils';
-import { AssetSearchWriter, AssetSearchWriterOptions } from '@/features/assets/search/asset-search.writer';
-import { FileSearchWriter, FileSearchWriterOptions } from '@/features/assets/search/file-search.writer';
-import { GeometryDetailRepo } from '@/features/geometries/geometry-detail.repo';
-import { GeometryRepo } from '@/features/geometries/geometry.repo';
-
-export const ASSET_ELASTIC_INDEX = 'swissgeol_asset_asset';
-export const FILE_ELASTIC_INDEX = 'swissgeol_asset_file';
-
-const SEARCH_BATCH_SIZE = 10_000;
-
-/**
- * Fields that are searchable via specific field queries (e.g. id:1234).
- * Note that these fields must be mapped as `keyword` in the Elasticsearch mapping (swissgeol_asset_asset.json).
- */
-const SEARCHABLE_FIELDS: (keyof ElasticsearchAsset)[] = [
-  'title',
-  'originalTitle',
-  'contactNames',
-  'sgsId',
-  'alternativeIds',
-  'id',
-];
+// Re-export constants and helpers for backwards compatibility
+export { ASSET_ELASTIC_INDEX, FILE_ELASTIC_INDEX } from '@/features/assets/search/asset-search.constants';
+export { escapeElasticQuery, normalizeFieldQuery } from '@/features/assets/search/asset-search.query';
 
 @Injectable()
 export class AssetSearchService {
-  private readonly logger = new Logger(AssetSearchService.name);
-
-  constructor(
-    private readonly elastic: ElasticsearchClient,
-    private readonly prisma: PrismaService,
-    private readonly assetRepo: AssetRepo,
-    private readonly geometryRepo: GeometryRepo,
-    private readonly geometryDetailRepo: GeometryDetailRepo,
-  ) {}
-
-  async register(asset: Asset): Promise<void> {
-    await Promise.all([this.getWriter().write(asset), this.getFileWriter().write(asset)]);
-  }
-
-  getWriter(options?: AssetSearchWriterOptions): AssetSearchWriter {
-    return new AssetSearchWriter(
-      this.elastic,
-      this.prisma,
-      this.geometryRepo,
-      this.geometryDetailRepo,
-      options ?? { index: ASSET_ELASTIC_INDEX, shouldRefresh: true },
-    );
-  }
-
-  getFileWriter(options?: FileSearchWriterOptions): FileSearchWriter {
-    return new FileSearchWriter(
-      this.elastic,
-      this.prisma,
-      this.geometryRepo,
-      options ?? { index: FILE_ELASTIC_INDEX, shouldRefresh: true },
-    );
-  }
-
-  async deleteFromIndex(assetId: number): Promise<void> {
-    await Promise.all([
-      this.elastic.delete({
-        index: ASSET_ELASTIC_INDEX,
-        id: `${assetId}`,
-        refresh: true,
-      }),
-      this.elastic.deleteByQuery({
-        index: FILE_ELASTIC_INDEX,
-        query: { term: { assetId: assetId } },
-        refresh: true,
-        ignore_unavailable: true,
-      }),
-    ]);
-  }
-
-  async count(): Promise<number> {
-    return (await this.elastic.count({ index: ASSET_ELASTIC_INDEX, ignore_unavailable: true })).count;
-  }
-
-  async countFiles(): Promise<number> {
-    return (await this.elastic.count({ index: FILE_ELASTIC_INDEX, ignore_unavailable: true })).count;
-  }
-
-  async syncWithDatabase(onProgress?: (percentage: number) => void | Promise<void>): Promise<void> {
-    // Write all Prisma assets into the sync index.
-    const total = await this.prisma.asset.count();
-    if (total === 0) {
-      this.logger.debug('No assets to sync');
-      if (onProgress != null) {
-        onProgress(1);
-      }
-      return;
-    }
-
-    // Initialize temporary sync indices.
-    const SYNC_INDEX = `sync-${ASSET_ELASTIC_INDEX}-${getDateTimeString()}`;
-    const FILE_SYNC_INDEX = `sync-${FILE_ELASTIC_INDEX}-${getDateTimeString()}`;
-
-    const existsSyncIndex = await this.elastic.indices.exists({ index: SYNC_INDEX });
-    if (existsSyncIndex) {
-      throw new Error(`can't sync to '${SYNC_INDEX}', index already exists`);
-    }
-    const existsFileSyncIndex = await this.elastic.indices.exists({ index: FILE_SYNC_INDEX });
-    if (existsFileSyncIndex) {
-      throw new Error(`can't sync to '${FILE_SYNC_INDEX}', index already exists`);
-    }
-
-    await this.elastic.indices.create({ index: SYNC_INDEX });
-    await this.elastic.indices.putMapping({ index: SYNC_INDEX, ...indexMapping });
-
-    await this.elastic.indices.create({ index: FILE_SYNC_INDEX });
-    await this.elastic.indices.putMapping({ index: FILE_SYNC_INDEX, ...fileIndexMapping });
-
-    const writer = this.getWriter({ index: SYNC_INDEX, isEager: true });
-    const fileWriter = this.getFileWriter({ index: FILE_SYNC_INDEX, isEager: true });
-    let offset = 0;
-    for (;;) {
-      this.logger.debug('Syncing assets.', {
-        total,
-        offset,
-        progress: Number((offset / total).toFixed(2)),
-      });
-      const records = await this.assetRepo.list({ limit: 1000, offset });
-      if (records.length === 0) {
-        break;
-      }
-      await Promise.all([writer.write(records), fileWriter.write(records)]);
-      offset += records.length;
-      if (onProgress != null) {
-        await onProgress(Math.min(offset / total, 1));
-      }
-    }
-    this.logger.debug('Done syncing assets.', { total });
-
-    // Delete existing indices and recreate them.
-    await this.elastic.indices.delete({ index: ASSET_ELASTIC_INDEX, ignore_unavailable: true });
-    await this.elastic.indices.create({ index: ASSET_ELASTIC_INDEX });
-    await this.elastic.indices.putMapping({ index: ASSET_ELASTIC_INDEX, ...indexMapping });
-
-    await this.elastic.indices.delete({ index: FILE_ELASTIC_INDEX, ignore_unavailable: true });
-    await this.elastic.indices.create({ index: FILE_ELASTIC_INDEX });
-    await this.elastic.indices.putMapping({ index: FILE_ELASTIC_INDEX, ...fileIndexMapping });
-
-    // Refresh and reindex sync indices into live indices.
-    await this.elastic.indices.refresh({ index: SYNC_INDEX });
-    await this.elastic.reindex({
-      source: { index: SYNC_INDEX },
-      dest: { index: ASSET_ELASTIC_INDEX },
-    });
-    await this.elastic.indices.refresh({ index: ASSET_ELASTIC_INDEX });
-    await this.elastic.indices.delete({ index: SYNC_INDEX });
-
-    await this.elastic.indices.refresh({ index: FILE_SYNC_INDEX });
-    await this.elastic.reindex({
-      source: { index: FILE_SYNC_INDEX },
-      dest: { index: FILE_ELASTIC_INDEX },
-    });
-    await this.elastic.indices.refresh({ index: FILE_ELASTIC_INDEX });
-    await this.elastic.indices.delete({ index: FILE_SYNC_INDEX });
-  }
+  constructor(private readonly elastic: ElasticsearchClient) {}
 
   /**
    * Searches for assets using a {@link AssetSearchQuery}.
@@ -241,82 +82,9 @@ export class AssetSearchService {
   }
 
   /**
-   * Searches for file pages matching the text query, with asset metadata filters applied.
-   * Returns file results with highlights, plus both asset and file total counts.
-   */
-  async searchFiles(
-    query: AssetSearchQuery,
-    user: User,
-    { limit = 100, offset = 0 }: PageOptions = {},
-  ): Promise<FileSearchResult> {
-    const elasticQuery = mapQueryToFileElasticDsl(query, user);
-
-    const response = await this.elastic.search({
-      index: FILE_ELASTIC_INDEX,
-      query: elasticQuery,
-      size: limit,
-      from: offset,
-      highlight: {
-        fields: {
-          content: {
-            fragment_size: 150,
-            number_of_fragments: 3,
-            pre_tags: ['<em>'],
-            post_tags: ['</em>'],
-          },
-        },
-      },
-      _source: ['fileId', 'assetId', 'assetTitle', 'fileName', 'page'],
-      track_total_hits: true,
-    });
-
-    const fileTotal = (response.hits.total as SearchTotalHits).value;
-
-    const data: FileSearchResultItem[] = response.hits.hits.map((hit) => {
-      const source = hit._source as Record<string, unknown>;
-      const highlights = hit.highlight?.['content'] ?? [];
-      return {
-        fileId: source['fileId'] as number,
-        assetId: source['assetId'] as number,
-        assetTitle: source['assetTitle'] as string,
-        fileName: source['fileName'] as string,
-        page: source['page'] as number,
-        highlights,
-      };
-    });
-
-    // Count matching assets in parallel.
-    const assetTotal = await this.countAssetsByQuery(query, user);
-
-    return {
-      page: {
-        offset,
-        size: data.length,
-        total: fileTotal,
-      },
-      assetTotal,
-      fileTotal,
-      data,
-    };
-  }
-
-  /**
-   * Count file pages matching the query in the file index.
-   */
-  async countFilesByQuery(query: AssetSearchQuery, user: User): Promise<number> {
-    const elasticQuery = mapQueryToFileElasticDsl(query, user);
-    const response = await this.elastic.count({
-      index: FILE_ELASTIC_INDEX,
-      query: elasticQuery,
-      ignore_unavailable: true,
-    });
-    return response.count;
-  }
-
-  /**
    * Count assets matching the query in the asset index.
    */
-  private async countAssetsByQuery(query: AssetSearchQuery, user: User): Promise<number> {
+  async countAssetsByQuery(query: AssetSearchQuery, user: User): Promise<number> {
     const elasticQuery = mapQueryToElasticDsl(query, user);
     const response = await this.elastic.count({
       index: ASSET_ELASTIC_INDEX,
@@ -471,142 +239,6 @@ export class AssetSearchService {
     };
   }
 
-  /**
-   * Aggregates the stats over all file pages matching a specific {@link AssetSearchQuery}.
-   * Runs on the file index, which has the same denormalized metadata fields as the asset index.
-   *
-   * @param query The query to match with.
-   * @param user The user that is executing the query.
-   */
-  async aggregateFiles(query: AssetSearchQuery, user: User): Promise<AssetSearchStats> {
-    type NestedAggResult<T> = {
-      [K in keyof T]: {
-        a: T[K];
-      };
-    };
-
-    interface Result {
-      minCreatedAt: { value: number };
-      maxCreatedAt: { value: number };
-      authorIds: {
-        buckets: AggregationBucket<ContactId>[];
-      };
-      kindCodes: {
-        buckets: AggregationBucket[];
-      };
-      languageCodes: {
-        buckets: AggregationBucket[];
-      };
-      geometryTypes: {
-        buckets: AggregationBucket<GeometryType | 'None'>[];
-      };
-      topicCodes: {
-        buckets: AggregationBucket[];
-      };
-      status: {
-        buckets: AggregationBucket<WorkflowStatus>[];
-      };
-      usageCodes: {
-        buckets: AggregationBucket<AssetSearchUsageCode>[];
-      };
-      workgroupIds: {
-        buckets: AggregationBucket<WorkgroupId>[];
-      };
-    }
-
-    interface AggregationBucket<K = string> {
-      key: K;
-      doc_count: number;
-    }
-
-    const makeAggregation = (
-      operator: 'terms' | 'min' | 'max',
-      groupName: string,
-      fieldName?: string,
-    ): AggregationsAggregationContainer => {
-      const NUMBER_OF_BUCKETS = 10_000;
-      const { filter } = mapQueryToFileElasticDslParts({ ...query, [groupName]: undefined }, user);
-      const field: { field: string; size?: number } = { field: fieldName ?? groupName, size: NUMBER_OF_BUCKETS };
-      if (operator !== 'terms') {
-        delete field.size;
-      }
-      return { aggs: { a: { [operator]: field } }, filter: { bool: { filter } } };
-    };
-
-    const { must, filter } = mapQueryToFileElasticDslParts(query, user);
-
-    const response = await this.elastic.search({
-      index: FILE_ELASTIC_INDEX,
-      size: 0,
-      query: { bool: { must, filter } },
-      track_total_hits: true,
-      ignore_unavailable: true,
-    });
-    const total = (response.hits.total as SearchTotalHits | undefined)?.value ?? 0;
-    if (total === 0) {
-      return {
-        total: 0,
-        kindCodes: [],
-        authorIds: [],
-        createdAt: null,
-        languageCodes: [],
-        geometryTypes: [],
-        topicCodes: [],
-        usageCodes: [],
-        workgroupIds: [],
-        status: [],
-      };
-    }
-
-    const aggregateByQuery = async (aggs: Record<string, AggregationsAggregationContainer>) => {
-      return await this.elastic.search({
-        index: FILE_ELASTIC_INDEX,
-        size: 0,
-        query: { bool: { must } },
-        track_total_hits: true,
-        aggregations: aggs,
-        filter_path: ['aggregations.*.a.buckets.*', 'aggregations.*.a.value'],
-        ignore_unavailable: true,
-      });
-    };
-
-    const result = await aggregateByQuery({
-      authorIds: makeAggregation('terms', 'authorIds'),
-      languageCodes: makeAggregation('terms', 'languageCodes'),
-      geometryTypes: makeAggregation('terms', 'geometryTypes'),
-      kindCodes: makeAggregation('terms', 'kindCodes', 'kindCode'),
-      topicCodes: makeAggregation('terms', 'topicCodes'),
-      usageCodes: makeAggregation('terms', 'usageCodes', 'usageCode'),
-      workgroupIds: makeAggregation('terms', 'workgroupIds', 'workgroupId'),
-      minCreatedAt: makeAggregation('min', 'minCreatedAt', 'createdAt'),
-      maxCreatedAt: makeAggregation('max', 'maxCreatedAt', 'createdAt'),
-      status: makeAggregation('terms', 'status', 'status'),
-    });
-
-    const aggs = result.aggregations as unknown as NestedAggResult<Result>;
-
-    const mapBucket = <T>(bucket: AggregationBucket<T>): ValueCount<T> => ({
-      value: bucket.key,
-      count: bucket.doc_count,
-    });
-
-    return {
-      total,
-      kindCodes: aggs.kindCodes?.a?.buckets?.map(mapBucket) ?? [],
-      authorIds: aggs.authorIds?.a?.buckets?.map(mapBucket) ?? [],
-      languageCodes: aggs.languageCodes?.a?.buckets?.map(mapBucket) ?? [],
-      geometryTypes: aggs.geometryTypes?.a?.buckets?.map(mapBucket) ?? [],
-      topicCodes: aggs.topicCodes?.a?.buckets?.map(mapBucket) ?? [],
-      usageCodes: aggs.usageCodes?.a?.buckets?.map(mapBucket) ?? [],
-      workgroupIds: aggs.workgroupIds?.a?.buckets?.map(mapBucket) ?? [],
-      createdAt: {
-        min: LocalDate.fromDate(new Date(aggs.minCreatedAt.a.value)),
-        max: LocalDate.fromDate(new Date(aggs.maxCreatedAt.a.value)),
-      },
-      status: aggs.status?.a?.buckets?.map(mapBucket) ?? [],
-    };
-  }
-
   private async searchAssetsByQuery(
     query: AssetSearchQuery,
     user: User,
@@ -705,7 +337,7 @@ export class AssetSearchService {
   }
 
   /**
-   * Execute an elastic query in the context of a @link SearchState}.
+   * Execute an elastic query in the context of a {@link SearchState}.
    *
    * This method will update {@link state.totalCount}.
    *
@@ -737,281 +369,4 @@ export class AssetSearchService {
     state.totalCount ??= (response.hits.total as SearchTotalHits).value as number;
     return response;
   }
-}
-
-const mapQueryToElasticDsl = (query: AssetSearchQuery, user: User): QueryDslQueryContainer => {
-  const { must, filter } = mapQueryToElasticDslParts(query, user);
-  return {
-    bool: {
-      must,
-      filter,
-    },
-  };
-};
-
-const mapQueryToElasticDslParts = (
-  query: AssetSearchQuery,
-  user: User,
-): { must: QueryDslQueryContainer[]; filter: QueryDslQueryContainer[] } => {
-  const scope = SEARCHABLE_FIELDS;
-  const queries: QueryDslQueryContainer[] = [];
-  const filters: QueryDslQueryContainer[] = [];
-
-  if (query.text != null && query.text.length > 0) {
-    queries.push({
-      bool: {
-        should: [
-          {
-            query_string: {
-              query: escapeElasticQuery(normalizeFieldQuery(query.text)),
-              fields: scope,
-            },
-          },
-        ],
-      },
-    });
-  }
-
-  if (query.authorId != null) {
-    filters.push({
-      term: {
-        authorIds: query.authorId,
-      },
-    });
-  }
-  if (query.createdAt != null && Object.keys(query.createdAt).length > 0) {
-    const createdAtFilter: QueryDslDateRangeQuery = {
-      format: 'yyyy-MM-dd',
-    };
-    if (query.createdAt.min != null) {
-      createdAtFilter.gte = query.createdAt.min.toString();
-    }
-    if (query.createdAt.max != null) {
-      createdAtFilter.lte = query.createdAt.max.toString();
-    }
-    filters.push({
-      range: {
-        createdAt: createdAtFilter,
-      },
-    });
-  }
-  if (query.topicCodes != null) {
-    filters.push(makeArrayFilter('topicCodes', query.topicCodes));
-  }
-  if (query.kindCodes != null) {
-    filters.push(makeArrayFilter('kindCode', query.kindCodes));
-  }
-  if (query.usageCodes != null) {
-    filters.push(makeArrayFilter('usageCode', query.usageCodes));
-  }
-  if (query.languageCodes != null) {
-    filters.push(makeArrayFilter('languageCodes', query.languageCodes));
-  }
-  if (query.geometryTypes != null) {
-    filters.push(makeArrayFilter('geometryTypes', query.geometryTypes));
-  }
-  if (query.status != null) {
-    filters.push(makeArrayFilter('status', query.status));
-  }
-  if (query.workgroupIds != null) {
-    filters.push({
-      terms: {
-        workgroupId: query.workgroupIds,
-      },
-    });
-  }
-  if (query.favoritesOnly) {
-    filters.push({
-      terms: {
-        favoredByUserIds: [user.id],
-      },
-    });
-  }
-
-  if (query.polygon != null) {
-    queries.push({
-      geo_polygon: {
-        locations: {
-          points: query.polygon.map(mapLv95ToElastic),
-        },
-      },
-    });
-  }
-  return {
-    must: queries,
-    filter: filters,
-  };
-};
-
-/**
- * Create an Elasticsearch filter that matches all documents which contain a specific field
- * that contains any of a given set of values (the _query_).
- *
- * When the query is empty, only documents for which the field does not exist are matched.
- * A field does also count as "not existing" when it is set to `null` or an empty array.
- *
- * @param field The field to match.
- * @param query The set of allowed values.
- */
-const makeArrayFilter = <T extends string | number>(field: string, query: T[]): QueryDslQueryContainer => {
-  if (query.length === 0) {
-    return { bool: { must_not: { exists: { field } } } };
-  }
-  return {
-    bool: {
-      should: query.map((term) => ({ term: { [field]: term } })),
-    },
-  };
-};
-
-/**
- * Maps the shared filter parts of an {@link AssetSearchQuery} to Elasticsearch filter clauses.
- * These filters apply to both asset and file indices (which share the same denormalized metadata fields).
- */
-const mapSharedFilterParts = (query: AssetSearchQuery, user: User): QueryDslQueryContainer[] => {
-  const filters: QueryDslQueryContainer[] = [];
-
-  if (query.authorId != null) {
-    filters.push({ term: { authorIds: query.authorId } });
-  }
-  if (query.createdAt != null && Object.keys(query.createdAt).length > 0) {
-    const createdAtFilter: QueryDslDateRangeQuery = { format: 'yyyy-MM-dd' };
-    if (query.createdAt.min != null) {
-      createdAtFilter.gte = query.createdAt.min.toString();
-    }
-    if (query.createdAt.max != null) {
-      createdAtFilter.lte = query.createdAt.max.toString();
-    }
-    filters.push({ range: { createdAt: createdAtFilter } });
-  }
-  if (query.topicCodes != null) {
-    filters.push(makeArrayFilter('topicCodes', query.topicCodes));
-  }
-  if (query.kindCodes != null) {
-    filters.push(makeArrayFilter('kindCode', query.kindCodes));
-  }
-  if (query.usageCodes != null) {
-    filters.push(makeArrayFilter('usageCode', query.usageCodes));
-  }
-  if (query.languageCodes != null) {
-    filters.push(makeArrayFilter('languageCodes', query.languageCodes));
-  }
-  if (query.geometryTypes != null) {
-    filters.push(makeArrayFilter('geometryTypes', query.geometryTypes));
-  }
-  if (query.status != null) {
-    filters.push(makeArrayFilter('status', query.status));
-  }
-  if (query.workgroupIds != null) {
-    filters.push({ terms: { workgroupId: query.workgroupIds } });
-  }
-  if (query.favoritesOnly) {
-    filters.push({ terms: { favoredByUserIds: [user.id] } });
-  }
-
-  return filters;
-};
-
-const mapQueryToFileElasticDsl = (query: AssetSearchQuery, user: User): QueryDslQueryContainer => {
-  const { must, filter } = mapQueryToFileElasticDslParts(query, user);
-  return { bool: { must, filter } };
-};
-
-const mapQueryToFileElasticDslParts = (
-  query: AssetSearchQuery,
-  user: User,
-): { must: QueryDslQueryContainer[]; filter: QueryDslQueryContainer[] } => {
-  const queries: QueryDslQueryContainer[] = [];
-  const filters = mapSharedFilterParts(query, user);
-
-  if (query.text != null && query.text.length > 0) {
-    queries.push({
-      match: {
-        content: {
-          query: query.text,
-          operator: 'and',
-        },
-      },
-    });
-  }
-
-  return { must: queries, filter: filters };
-};
-
-const getDateTimeString = (): string => {
-  const now = new Date();
-  const padZero = (value: number): string | number => {
-    if (value >= 10) {
-      return value;
-    }
-    return `0${value}`;
-  };
-  return (
-    '' +
-    now.getUTCFullYear() +
-    padZero(now.getUTCMonth()) +
-    padZero(now.getUTCDate()) +
-    padZero(now.getUTCHours()) +
-    padZero(now.getUTCMinutes()) +
-    padZero(now.getUTCSeconds())
-  );
-};
-
-interface PageOptions {
-  limit?: number;
-  offset?: number;
-}
-
-/**
- * Escapes special characters in an Elasticsearch query string. Preserves wildcards (*) since this is what users might
- * use. This is a fix for a long-standing issue after io-ts refactor (see assets-639 for details).
- *
- * Furthermore, it also escapes the colon (:) character, but then un-escapes it for specific fields that are allowed to
- * be searched, to avoid issues when users search for e.g. "title:My Title".
- */
-export const escapeElasticQuery = (query: string): string => {
-  let escaped = query.replace(/(&&|\|\||!|\(|\)|\{|}|\[|]|\^|"|~|\+|-|=|\?|:|\\|\/)/g, '\\$1');
-
-  for (const field of SEARCHABLE_FIELDS) {
-    escaped = escaped.replaceAll(new RegExp(String.raw`\b${field}\\:`, 'gi'), `${field}:`);
-  }
-
-  return escaped;
-};
-
-/**
- * Helper function for backwards compatibility of field names in search queries.
- */
-export const normalizeFieldQuery = (query: string): string =>
-  query
-    .replace(/title(_*)public:/gi, 'title:')
-    .replace(/title(_*)original:/gi, 'originalTitle:')
-    .replace(/contact(_*)name:/gi, 'contactNames:')
-    .replace(/asset(_*)id:/gi, 'id:')
-    .replace(/sgs(_*)id:/gi, 'sgsId:');
-
-/**
- * The state of a multistep asset search.
- */
-interface SearchState {
-  /**
-   * The assets that match the search.
-   * This is a mapping from the assets' id to their serialized JSON string.
-   */
-  matchedAssets: Map<AssetId, AssetJSON>;
-
-  /**
-   * The id of the last asset that has been matched.
-   * This is used to enable paginated search results with Elasticsearch.
-   *
-   * This is `null` if no query has been executed yet.
-   */
-  lastAssetId: number | null;
-
-  /**
-   * The total number of assets, including the ones that were skipped due to offset and limit.
-   *
-   * This is `null` if no query has been executed yet.
-   */
-  totalCount: number | null;
 }
